quality/requirements.txt
absl-py==2.1.0
accelerate==1.3.0
accelerator==2024.9.13
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiolimiter==1.2.1
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anthropic==0.52.2
antlr4-python3-runtime==4.7.2
anyio==4.8.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.1.0
blake3==1.0.4
bottle==0.12.25
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
compressed-tensors==0.9.3
contourpy==1.3.1
cupy-cuda12x==13.4.1
cycler==0.12.1
dataproperty==1.1.0
datasets==3.2.0
decord==0.6.0
deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
editdistance==0.8.1
einops==0.8.1
email-validator==2.2.0
evaluate==0.4.3
fastapi==0.115.9
fastapi-cli==0.0.7
fastrlock==0.8.3
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
fuzzywuzzy==0.18.0
gguf==0.16.3
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.14.0
hf-xet==1.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
immutabledict==4.2.1
importlib-metadata==8.0.0
interegular==0.3.3
jieba==0.42.1
jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langdetect==1.0.9
lark==1.2.2
latex2sympy2==1.9.1
llguidance==0.7.19
llvmlite==0.44.0
lm-eval==0.3.0
lm-format-enforcer==0.10.11
lxml==5.3.1
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.10.0
mbstrdecoder==1.1.4
mdurl==0.1.2
mistral-common==1.5.4
modelscope==1.23.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.2
numexpr==2.10.2
numpy==2.2.2
openai==1.63.0
orjson==3.10.15
outlines==0.1.11
outlines-core==0.1.26
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
pathvalidate==3.2.3
peft==0.14.0
pillow==11.1.0
plotly==5.24.1
ply==3.11
portalocker==3.1.1
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.1.0
propcache==0.2.1
protobuf==4.25.7
py-cpuinfo==9.0.0
pyarrow==19.0.0
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.10.6
pydantic-core==2.27.2
pydub==0.25.1
pympler==1.1
pyparsing==3.2.1
pysubs2==1.8.0
pytablewriter==1.2.1
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.1
pyyaml==6.0.2
qwen-vl-utils==0.0.11
ray==2.46.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.14.5
rouge-chinese==1.0.3
rouge-score==0.1.2
rpds-py==0.24.0
ruff==0.9.8
sacrebleu==1.5.0
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
setproctitle==1.3.4
setuptools==69.5.1
shellingham==1.5.4
simple-ddl-parser==1.7.1
simplejson==3.20.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlitedict==2.1.0
starlette==0.45.3
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
tomlkit==0.12.0
torchaudio==2.6.0
torchcodec
torchvision==0.21.0
tqdm==4.67.1
tqdm-multiprocess==0.0.11
transformers==4.51.3
transformers-stream-generator==0.0.5
triton==3.2.0
typepy==1.3.4
typer==0.15.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vllm==0.8.5.post1
waitress==3.0.2
watchfiles==1.0.5
websockets==12.0
word2number==1.1
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
